==> source/Watch/code/1. crawling-movie.ipynb <==
# 영화
영화에 관련된 데이터를 네이버 영화, 왓챠에서 크롤링해온다

import requests
import pandas as pd
import time
import csv
from bs4 import BeautifulSoup
from selenium import webdriver
from tqdm.notebook import tqdm

## 왓챠

왓챠에서 '김지빈' 유저가 평가한 영화 목록을 가져온다

url = 'https://watcha.com/ko-KR/users/24BqE2RY8vrjn/contents/movies/ratings'
driver = webdriver.Chrome('./chromedriver')
driver.implicitly_wait(3)

driver.get(url)

SCROLL_PAUSE_TIME=2.5
last_height = driver.execute_script("return document.body.scrollHeight")

scroll_num = 0

while True:
    driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")

    time.sleep(SCROLL_PAUSE_TIME)
    driver.execute_script("window.scrollTo(0, document.body.scrollHeight-50);")
    time.sleep(SCROLL_PAUSE_TIME)

    new_height = driver.execute_script("return document.body.scrollHeight")

    if new_height == last_height:
        break
    
    scroll_num += 1
    last_height = new_height

page = driver.page_source
soup = BeautifulSoup(page, 'html.parser')

movie_list=soup.select('li.css-106b4k6-Self.e3fgkal0 > a')
movies=["https://watcha.com"+movie_list[i].get('href') for i in range(len(movie_list))]

print(movies)
with open('movie_list_raw.txt', 'w') as file:
    file.write('\n'.join(movies)) #read -> file.readlines()

movies=['https://watcha.com/ko-KR/contents/mQO8625', 'https://watcha.com/ko-KR/contents/mWyJbrM', 'https://watcha.com/ko-KR/contents/mOk6r9z', 'https://watcha.com/ko-KR/contents/mdB7yP5', 'https://watcha.com/ko-KR/contents/mO1Vx4Y', 'https://watcha.com/ko-KR/contents/mAd7lv5', 'https://watcha.com/ko-KR/contents/mWywz9V', 'https://watcha.com/ko-KR/contents/mdKD2aV', 'https://watcha.com/ko-KR/contents/mWyJa0K', 'https://watcha.com/ko-KR/contents/m5ZlXxL', 'https://watcha.com/ko-KR/contents/mWpJvAd', 'https://watcha.com/ko-KR/contents/mOAZ6Rl', 'https://watcha.com/ko-KR/contents/mOPNXw5', 'https://watcha.com/ko-KR/contents/mOPomg7', 'https://watcha.com/ko-KR/contents/m5Q9Zg8', 'https://watcha.com/ko-KR/contents/md7r0l1', 'https://watcha.com/ko-KR/contents/m5NVKVY', 'https://watcha.com/ko-KR/contents/mOP0v0Y', 'https://watcha.com/ko-KR/contents/mdErj22', 'https://watcha.com/ko-KR/contents/m5ZZ2P9', 'https://watcha.com/ko-KR/contents/m5nQrpo', 'https://watcha.com/ko-KR/contents/mOVvVyp', 'https://watcha.com/ko-KR/contents/mdKBgy9', 'https://watcha.com/ko-KR/contents/md76YRp', 'https://watcha.com/ko-KR/contents/mdRL4eL', 'https://watcha.com/ko-KR/contents/m5ewr3z', 'https://watcha.com/ko-KR/contents/m5XM27B', 'https://watcha.com/ko-KR/contents/mOVDa1p', 'https://watcha.com/ko-KR/contents/mdRLXvW', 'https://watcha.com/ko-KR/contents/mWvkVBx', 'https://watcha.com/ko-KR/contents/mdBzNnP', 'https://watcha.com/ko-KR/contents/mdjwR8l', 'https://watcha.com/ko-KR/contents/mOgoNMO', 'https://watcha.com/ko-KR/contents/mWyJxx1', 'https://watcha.com/ko-KR/contents/mWwlN3O', 'https://watcha.com/ko-KR/contents/m533xYy', 'https://watcha.com/ko-KR/contents/mWqJbZX', 'https://watcha.com/ko-KR/contents/mO0gQ41', 'https://watcha.com/ko-KR/contents/mdRLZVK', 'https://watcha.com/ko-KR/contents/md76rkk', 'https://watcha.com/ko-KR/contents/mLOPBa5', 'https://watcha.com/ko-KR/contents/mY5QZDW', 'https://watcha.com/ko-KR/contents/m5nQwly', 'https://watcha.com/ko-KR/contents/mWJyX7x', 'https://watcha.com/ko-KR/contents/m5xJZE3', 'https://watcha.com/ko-KR/contents/mObGV9n', 'https://watcha.com/ko-KR/contents/mWJyDKM', 'https://watcha.com/ko-KR/contents/mWqJy9K', 'https://watcha.com/ko-KR/contents/m5m1YNG', 'https://watcha.com/ko-KR/contents/mOVv2Lz', 'https://watcha.com/ko-KR/contents/m5awllO', 'https://watcha.com/ko-KR/contents/mOAZkEG', 'https://watcha.com/ko-KR/contents/mdKBpoY', 'https://watcha.com/ko-KR/contents/mdRLbZr', 'https://watcha.com/ko-KR/contents/mWyJa2k', 'https://watcha.com/ko-KR/contents/mWyJYkO', 'https://watcha.com/ko-KR/contents/m5GXo8r', 'https://watcha.com/ko-KR/contents/m5QqAPG', 'https://watcha.com/ko-KR/contents/mdjawmz', 'https://watcha.com/ko-KR/contents/mOVvDjg', 'https://watcha.com/ko-KR/contents/m5Y6yQW', 'https://watcha.com/ko-KR/contents/md7Z6vO', 'https://watcha.com/ko-KR/contents/mdMgNAJ', 'https://watcha.com/ko-KR/contents/mO8KXmX', 'https://watcha.com/ko-KR/contents/m5Y6LNl', 'https://watcha.com/ko-KR/contents/mdEwMrg', 'https://watcha.com/ko-KR/contents/mWyJ8VO', 'https://watcha.com/ko-KR/contents/mWzw6E0', 'https://watcha.com/ko-KR/contents/mOAZkKm', 'https://watcha.com/ko-KR/contents/m5xDb35', 'https://watcha.com/ko-KR/contents/mdRL434', 'https://watcha.com/ko-KR/contents/mOPoAXW', 'https://watcha.com/ko-KR/contents/m5xJxeB', 'https://watcha.com/ko-KR/contents/mWLjDm6', 'https://watcha.com/ko-KR/contents/m5xDab5', 'https://watcha.com/ko-KR/contents/mWvkxbr', 'https://watcha.com/ko-KR/contents/mWyJwbr', 'https://watcha.com/ko-KR/contents/m5DGQDQ', 'https://watcha.com/ko-KR/contents/mdMlb4O', 'https://watcha.com/ko-KR/contents/mOgBBKY', 'https://watcha.com/ko-KR/contents/md6X2aW', 'https://watcha.com/ko-KR/contents/mOVvJ68', 'https://watcha.com/ko-KR/contents/mW4LYVL', 'https://watcha.com/ko-KR/contents/m5QqQ77', 'https://watcha.com/ko-KR/contents/m5m1KM7', 'https://watcha.com/ko-KR/contents/mOl4GK6', 'https://watcha.com/ko-KR/contents/m5GXDep', 'https://watcha.com/ko-KR/contents/mWyJD46', 'https://watcha.com/ko-KR/contents/mO1VKeL', 'https://watcha.com/ko-KR/contents/mOk6oAk', 'https://watcha.com/ko-KR/contents/m5NVpDp', 'https://watcha.com/ko-KR/contents/m5GXzED', 'https://watcha.com/ko-KR/contents/mWwBw83', 'https://watcha.com/ko-KR/contents/mOl4x7g', 'https://watcha.com/ko-KR/contents/mWvkKME', 'https://watcha.com/ko-KR/contents/mWwBwA3', 'https://watcha.com/ko-KR/contents/mWzw7oZ', 'https://watcha.com/ko-KR/contents/mWvkXvZ', 'https://watcha.com/ko-KR/contents/mWyg9kd', 'https://watcha.com/ko-KR/contents/m5eJD2R', 'https://watcha.com/ko-KR/contents/mWpmGjx', 'https://watcha.com/ko-KR/contents/mWLjMPr', 'https://watcha.com/ko-KR/contents/mMO2zkO', 'https://watcha.com/ko-KR/contents/mWywaMd', 'https://watcha.com/ko-KR/contents/mObqb7W', 'https://watcha.com/ko-KR/contents/mW4L6jY', 'https://watcha.com/ko-KR/contents/m5aVMGQ', 'https://watcha.com/ko-KR/contents/mr5ZLlO', 'https://watcha.com/ko-KR/contents/mWw8zEO', 'https://watcha.com/ko-KR/contents/m533vQ2', 'https://watcha.com/ko-KR/contents/mOgBnv0', 'https://watcha.com/ko-KR/contents/m5QKK8O', 'https://watcha.com/ko-KR/contents/mOgb1NW', 'https://watcha.com/ko-KR/contents/mWqJeDX', 'https://watcha.com/ko-KR/contents/md6YvoA', 'https://watcha.com/ko-KR/contents/mO2MNRn', 'https://watcha.com/ko-KR/contents/mO2MNlx', 'https://watcha.com/ko-KR/contents/mWwBb4q', 'https://watcha.com/ko-KR/contents/mgdMnm5', 'https://watcha.com/ko-KR/contents/m5aV96M', 'https://watcha.com/ko-KR/contents/mW93v7e', 'https://watcha.com/ko-KR/contents/mdBzrpD', 'https://watcha.com/ko-KR/contents/mWJypq8', 'https://watcha.com/ko-KR/contents/m5GX0zB', 'https://watcha.com/ko-KR/contents/m5ZXblx', 'https://watcha.com/ko-KR/contents/m5Y6QqQ', 'https://watcha.com/ko-KR/contents/mW93vGe', 'https://watcha.com/ko-KR/contents/mOVvVxY', 'https://watcha.com/ko-KR/contents/mOk60yQ', 'https://watcha.com/ko-KR/contents/mOo0DqZ', 'https://watcha.com/ko-KR/contents/m5NVMAX', 'https://watcha.com/ko-KR/contents/mOg0eNd', 'https://watcha.com/ko-KR/contents/m5eJ1wR', 'https://watcha.com/ko-KR/contents/mObG11N', 'https://watcha.com/ko-KR/contents/mQO8b5R', 'https://watcha.com/ko-KR/contents/m5xJNzl', 'https://watcha.com/ko-KR/contents/m533Nnn', 'https://watcha.com/ko-KR/contents/mOk6ENg', 'https://watcha.com/ko-KR/contents/m5aV9XM', 'https://watcha.com/ko-KR/contents/mOPomng', 'https://watcha.com/ko-KR/contents/m5m1XlB', 'https://watcha.com/ko-KR/contents/mO8KDxX', 'https://watcha.com/ko-KR/contents/mOgBozg', 'https://watcha.com/ko-KR/contents/mOVyqqW', 'https://watcha.com/ko-KR/contents/mdBzP4A', 'https://watcha.com/ko-KR/contents/mObG9n7', 'https://watcha.com/ko-KR/contents/mO8KywG', 'https://watcha.com/ko-KR/contents/mWqJbLo', 'https://watcha.com/ko-KR/contents/m5NVjDa', 'https://watcha.com/ko-KR/contents/md6YDDq', 'https://watcha.com/ko-KR/contents/m5Qq28G', 'https://watcha.com/ko-KR/contents/m5nQD4X', 'https://watcha.com/ko-KR/contents/mWLjM30', 'https://watcha.com/ko-KR/contents/m85XRAO', 'https://watcha.com/ko-KR/contents/mWqJbYB', 'https://watcha.com/ko-KR/contents/mdBzr12', 'https://watcha.com/ko-KR/contents/mWwBMP3', 'https://watcha.com/ko-KR/contents/mW4LxAo', 'https://watcha.com/ko-KR/contents/mOVaaRW', 'https://watcha.com/ko-KR/contents/mObGM6z', 'https://watcha.com/ko-KR/contents/md76JgM', 'https://watcha.com/ko-KR/contents/mO8KDEX', 'https://watcha.com/ko-KR/contents/md71l1W', 'https://watcha.com/ko-KR/contents/mWqJyRK', 'https://watcha.com/ko-KR/contents/mWyJqvy', 'https://watcha.com/ko-KR/contents/md6YDVa', 'https://watcha.com/ko-KR/contents/mObGr9R', 'https://watcha.com/ko-KR/contents/mWLVrpd', 'https://watcha.com/ko-KR/contents/mDWvjEW', 'https://watcha.com/ko-KR/contents/md76J08', 'https://watcha.com/ko-KR/contents/m5ZX9ke', 'https://watcha.com/ko-KR/contents/mdMBAR5', 'https://watcha.com/ko-KR/contents/mOAZjpG', 'https://watcha.com/ko-KR/contents/mdEwMvg', 'https://watcha.com/ko-KR/contents/m5QqpLQ', 'https://watcha.com/ko-KR/contents/mOAZ6Pd', 'https://watcha.com/ko-KR/contents/mOo0PKZ', 'https://watcha.com/ko-KR/contents/mWqJboK', 'https://watcha.com/ko-KR/contents/mMO2GdK', 'https://watcha.com/ko-KR/contents/mWzxGDd', 'https://watcha.com/ko-KR/contents/m5m1v63', 'https://watcha.com/ko-KR/contents/md76DgB', 'https://watcha.com/ko-KR/contents/m5XMMN3', 'https://watcha.com/ko-KR/contents/mdRJqLd', 'https://watcha.com/ko-KR/contents/m533L7p', 'https://watcha.com/ko-KR/contents/mWLjk86', 'https://watcha.com/ko-KR/contents/md76ze8', 'https://watcha.com/ko-KR/contents/m5eJyM9', 'https://watcha.com/ko-KR/contents/m5m1vJx', 'https://watcha.com/ko-KR/contents/m5aVlPj', 'https://watcha.com/ko-KR/contents/myWqxp5', 'https://watcha.com/ko-KR/contents/m5XMqmD', 'https://watcha.com/ko-KR/contents/mOl4YaQ', 'https://watcha.com/ko-KR/contents/mW93opG', 'https://watcha.com/ko-KR/contents/mO1VzlQ', 'https://watcha.com/ko-KR/contents/m5m1vy7', 'https://watcha.com/ko-KR/contents/mdRLwlv', 'https://watcha.com/ko-KR/contents/mpWpZ6d', 'https://watcha.com/ko-KR/contents/mBOkzJW', 'https://watcha.com/ko-KR/contents/mgOAPWm', 'https://watcha.com/ko-KR/contents/md6YxNy', 'https://watcha.com/ko-KR/contents/m5XMv3W', 'https://watcha.com/ko-KR/contents/m5Y6pDl', 'https://watcha.com/ko-KR/contents/mWwBZ9G', 'https://watcha.com/ko-KR/contents/m5xJk3x', 'https://watcha.com/ko-KR/contents/mW93y9B', 'https://watcha.com/ko-KR/contents/mW9EZlW', 'https://watcha.com/ko-KR/contents/m5eJZ3R', 'https://watcha.com/ko-KR/contents/m5YMvR5', 'https://watcha.com/ko-KR/contents/m5GXz42', 'https://watcha.com/ko-KR/contents/mOgBo03', 'https://watcha.com/ko-KR/contents/mOAZ86G', 'https://watcha.com/ko-KR/contents/my5YJEW', 'https://watcha.com/ko-KR/contents/mWJyPpx', 'https://watcha.com/ko-KR/contents/mW93GDZ', 'https://watcha.com/ko-KR/contents/m5n4oYO', 'https://watcha.com/ko-KR/contents/m5rQxJE', 'https://watcha.com/ko-KR/contents/mdMgK6w', 'https://watcha.com/ko-KR/contents/md66PXd', 'https://watcha.com/ko-KR/contents/mWvkX7e', 'https://watcha.com/ko-KR/contents/mWJyL1Q', 'https://watcha.com/ko-KR/contents/mO1V3ok', 'https://watcha.com/ko-KR/contents/m5xJMDr', 'https://watcha.com/ko-KR/contents/mO0gZP1', 'https://watcha.com/ko-KR/contents/mWyJ77M', 'https://watcha.com/ko-KR/contents/mdBzV9D', 'https://watcha.com/ko-KR/contents/mW4LkqY', 'https://watcha.com/ko-KR/contents/mWvkyor', 'https://watcha.com/ko-KR/contents/mO8K0Xb', 'https://watcha.com/ko-KR/contents/m6dRVad', 'https://watcha.com/ko-KR/contents/mW4Lzqg', 'https://watcha.com/ko-KR/contents/m5x6KlO', 'https://watcha.com/ko-KR/contents/mOAZ89E', 'https://watcha.com/ko-KR/contents/mmObrNO', 'https://watcha.com/ko-KR/contents/m5xJXvb', 'https://watcha.com/ko-KR/contents/m5GV6Yd', 'https://watcha.com/ko-KR/contents/m533Z8p', 'https://watcha.com/ko-KR/contents/mdjaLA9', 'https://watcha.com/ko-KR/contents/m45nnX5', 'https://watcha.com/ko-KR/contents/mv5axqd', 'https://watcha.com/ko-KR/contents/mjO0ovW', 'https://watcha.com/ko-KR/contents/m5ZXBo0', 'https://watcha.com/ko-KR/contents/m5xDyj5', 'https://watcha.com/ko-KR/contents/m5aVlap', 'https://watcha.com/ko-KR/contents/m9dBqzO', 'https://watcha.com/ko-KR/contents/mWzwVq9', 'https://watcha.com/ko-KR/contents/mOPoLaP', 'https://watcha.com/ko-KR/contents/mWJykq5', 'https://watcha.com/ko-KR/contents/mdja0A9', 'https://watcha.com/ko-KR/contents/mOl48qL', 'https://watcha.com/ko-KR/contents/mWq9waW', 'https://watcha.com/ko-KR/contents/m5xJngq', 'https://watcha.com/ko-KR/contents/mWLRRDd', 'https://watcha.com/ko-KR/contents/mdjaQyl', 'https://watcha.com/ko-KR/contents/mdEwA1p', 'https://watcha.com/ko-KR/contents/mO8KzlX', 'https://watcha.com/ko-KR/contents/md7DkB5', 'https://watcha.com/ko-KR/contents/mO2M33G', 'https://watcha.com/ko-KR/contents/mNdKveO', 'https://watcha.com/ko-KR/contents/mWLjVZD', 'https://watcha.com/ko-KR/contents/mWJmgx5', 'https://watcha.com/ko-KR/contents/mOl44bQ', 'https://watcha.com/ko-KR/contents/mdMwwJW', 'https://watcha.com/ko-KR/contents/mOAVmGd', 'https://watcha.com/ko-KR/contents/mW4La4l', 'https://watcha.com/ko-KR/contents/mWvk10e', 'https://watcha.com/ko-KR/contents/m5rQQg7', 'https://watcha.com/ko-KR/contents/mO2MKGW', 'https://watcha.com/ko-KR/contents/mdBzQED', 'https://watcha.com/ko-KR/contents/mdEwAR2', 'https://watcha.com/ko-KR/contents/mdEwxpW', 'https://watcha.com/ko-KR/contents/m5QZ8QW', 'https://watcha.com/ko-KR/contents/mOgBxPa', 'https://watcha.com/ko-KR/contents/mOVv6zq', 'https://watcha.com/ko-KR/contents/m5xJLaO', 'https://watcha.com/ko-KR/contents/m5DGE4B', 'https://watcha.com/ko-KR/contents/md7x8N5', 'https://watcha.com/ko-KR/contents/mWqJMnp', 'https://watcha.com/ko-KR/contents/mW9xB85', 'https://watcha.com/ko-KR/contents/m5alkNd', 'https://watcha.com/ko-KR/contents/md768n1', 'https://watcha.com/ko-KR/contents/mOVlo85', 'https://watcha.com/ko-KR/contents/mgdMr45', 'https://watcha.com/ko-KR/contents/m5Y6lKG', 'https://watcha.com/ko-KR/contents/m5nQ4Dd', 'https://watcha.com/ko-KR/contents/mnWJnqW', 'https://watcha.com/ko-KR/contents/my5YvvW', 'https://watcha.com/ko-KR/contents/mWJjpQW', 'https://watcha.com/ko-KR/contents/m5DpxB5', 'https://watcha.com/ko-KR/contents/mWqeQrW', 'https://watcha.com/ko-KR/contents/mWqJvBO', 'https://watcha.com/ko-KR/contents/m5Zx6LO', 'https://watcha.com/ko-KR/contents/mO2lEQ5', 'https://watcha.com/ko-KR/contents/mOVvgo8', 'https://watcha.com/ko-KR/contents/mOAZPad', 'https://watcha.com/ko-KR/contents/m5aNrDW', 'https://watcha.com/ko-KR/contents/mO8XobO', 'https://watcha.com/ko-KR/contents/mpWpow5', 'https://watcha.com/ko-KR/contents/mP5mAx5', 'https://watcha.com/ko-KR/contents/mdEkoAW', 'https://watcha.com/ko-KR/contents/mMO2JAO', 'https://watcha.com/ko-KR/contents/mWpmax4', 'https://watcha.com/ko-KR/contents/mWwVo45', 'https://watcha.com/ko-KR/contents/md7xGv5', 'https://watcha.com/ko-KR/contents/mLOP7lW', 'https://watcha.com/ko-KR/contents/mO8KR42', 'https://watcha.com/ko-KR/contents/mP5m7jO', 'https://watcha.com/ko-KR/contents/mOl4nEg', 'https://watcha.com/ko-KR/contents/m5DGg3d', 'https://watcha.com/ko-KR/contents/m5XMJEA', 'https://watcha.com/ko-KR/contents/mdBzm6G', 'https://watcha.com/ko-KR/contents/mP5mj7W', 'https://watcha.com/ko-KR/contents/m5DVDDO', 'https://watcha.com/ko-KR/contents/mAOl2L5', 'https://watcha.com/ko-KR/contents/mdjagwO', 'https://watcha.com/ko-KR/contents/m5QqYD5', 'https://watcha.com/ko-KR/contents/mOVvJYO', 'https://watcha.com/ko-KR/contents/m5rQG1E', 'https://watcha.com/ko-KR/contents/m5m14vD', 'https://watcha.com/ko-KR/contents/mW939lK', 'https://watcha.com/ko-KR/contents/mWqJnGr', 'https://watcha.com/ko-KR/contents/mWzwV11', 'https://watcha.com/ko-KR/contents/m5mQx35', 'https://watcha.com/ko-KR/contents/mdB61Ad', 'https://watcha.com/ko-KR/contents/m5XML4L', 'https://watcha.com/ko-KR/contents/mY5QNRW', 'https://watcha.com/ko-KR/contents/m5Y0qRd', 'https://watcha.com/ko-KR/contents/mBOkGQO', 'https://watcha.com/ko-KR/contents/mdRLw3r', 'https://watcha.com/ko-KR/contents/mWJJGxW', 'https://watcha.com/ko-KR/contents/mO2MK0n', 'https://watcha.com/ko-KR/contents/mAOlnl5', 'https://watcha.com/ko-KR/contents/m5YrbQW', 'https://watcha.com/ko-KR/contents/mOkXVQd', 'https://watcha.com/ko-KR/contents/md76wkO', 'https://watcha.com/ko-KR/contents/mOA9gEO', 'https://watcha.com/ko-KR/contents/mO8GEjW', 'https://watcha.com/ko-KR/contents/mWpme4M', 'https://watcha.com/ko-KR/contents/mJ5xpq5', 'https://watcha.com/ko-KR/contents/mdjabNO', 'https://watcha.com/ko-KR/contents/mJOVPq5', 'https://watcha.com/ko-KR/contents/mXOgez5', 'https://watcha.com/ko-KR/contents/mYdEogW', 'https://watcha.com/ko-KR/contents/mW4La3e', 'https://watcha.com/ko-KR/contents/md76Zb8', 'https://watcha.com/ko-KR/contents/mdBzV4k', 'https://watcha.com/ko-KR/contents/mdRLx0V', 'https://watcha.com/ko-KR/contents/mWy1NkW', 'https://watcha.com/ko-KR/contents/mAd738d', 'https://watcha.com/ko-KR/contents/m5eJx3g', 'https://watcha.com/ko-KR/contents/mWwBVzq', 'https://watcha.com/ko-KR/contents/m5DGD23', 'https://watcha.com/ko-KR/contents/m5m1lg5', 'https://watcha.com/ko-KR/contents/md76nkO', 'https://watcha.com/ko-KR/contents/mOVv611', 'https://watcha.com/ko-KR/contents/m5NVVwR', 'https://watcha.com/ko-KR/contents/md6xEAW', 'https://watcha.com/ko-KR/contents/mOAZ6Zm', 'https://watcha.com/ko-KR/contents/m5n07VO', 'https://watcha.com/ko-KR/contents/md6k3q5', 'https://watcha.com/ko-KR/contents/mOo0JZd', 'https://watcha.com/ko-KR/contents/mE5D435', 'https://watcha.com/ko-KR/contents/mdEkbEW', 'https://watcha.com/ko-KR/contents/mOgBwNW', 'https://watcha.com/ko-KR/contents/mOg8BMd', 'https://watcha.com/ko-KR/contents/mwWLm0O', 'https://watcha.com/ko-KR/contents/mOkEMz5', 'https://watcha.com/ko-KR/contents/mdjagDO', 'https://watcha.com/ko-KR/contents/mdK2Q9O', 'https://watcha.com/ko-KR/contents/mdKvL7O', 'https://watcha.com/ko-KR/contents/mWJyQM5', 'https://watcha.com/ko-KR/contents/mObGwnd', 'https://watcha.com/ko-KR/contents/mP5mYm5', 'https://watcha.com/ko-KR/contents/mLOPXO6', 'https://watcha.com/ko-KR/contents/mAd7415', 'https://watcha.com/ko-KR/contents/m5nZDV5', 'https://watcha.com/ko-KR/contents/mO2MQNW', 'https://watcha.com/ko-KR/contents/m85XqAW', 'https://watcha.com/ko-KR/contents/mOgBVzW', 'https://watcha.com/ko-KR/contents/mgdM6k5', 'https://watcha.com/ko-KR/contents/mJOV1RW', 'https://watcha.com/ko-KR/contents/mY5Qlyd', 'https://watcha.com/ko-KR/contents/mM5rEdq', 'https://watcha.com/ko-KR/contents/mO2MDxW', 'https://watcha.com/ko-KR/contents/mOPBwV5', 'https://watcha.com/ko-KR/contents/mOVvApO', 'https://watcha.com/ko-KR/contents/mO0gDVW', 'https://watcha.com/ko-KR/contents/mAd7YNW', 'https://watcha.com/ko-KR/contents/mWJarx5', 'https://watcha.com/ko-KR/contents/mBOk88d', 'https://watcha.com/ko-KR/contents/mjO0PzO', 'https://watcha.com/ko-KR/contents/mdEwQ2W', 'https://watcha.com/ko-KR/contents/mWpKxx5', 'https://watcha.com/ko-KR/contents/mZ5e9R5', 'https://watcha.com/ko-KR/contents/m5NMpXd', 'https://watcha.com/ko-KR/contents/mO0rl15', 'https://watcha.com/ko-KR/contents/mW4rmed', 'https://watcha.com/ko-KR/contents/mdMgLx2', 'https://watcha.com/ko-KR/contents/mWqJ28D', 'https://watcha.com/ko-KR/contents/mO2MKxA', 'https://watcha.com/ko-KR/contents/mWLjR9k', 'https://watcha.com/ko-KR/contents/m5NYXpd', 'https://watcha.com/ko-KR/contents/m5rQoaV', 'https://watcha.com/ko-KR/contents/mOoJKyW', 'https://watcha.com/ko-KR/contents/m5NVEE2', 'https://watcha.com/ko-KR/contents/mO8KKXD', 'https://watcha.com/ko-KR/contents/mOPoAqa', 'https://watcha.com/ko-KR/contents/mO0g2Yx', 'https://watcha.com/ko-KR/contents/mOPo6Dl', 'https://watcha.com/ko-KR/contents/mWpkgM5', 'https://watcha.com/ko-KR/contents/mObGGj7', 'https://watcha.com/ko-KR/contents/m5n31DW', 'https://watcha.com/ko-KR/contents/m5Gk725', 'https://watcha.com/ko-KR/contents/mdK3Qkd', 'https://watcha.com/ko-KR/contents/mdReBLO', 'https://watcha.com/ko-KR/contents/mOANbld', 'https://watcha.com/ko-KR/contents/mWyP8Y5', 'https://watcha.com/ko-KR/contents/m5ZXX2e', 'https://watcha.com/ko-KR/contents/m5QqDl5', 'https://watcha.com/ko-KR/contents/md76veO', 'https://watcha.com/ko-KR/contents/mOP7XpW', 'https://watcha.com/ko-KR/contents/m5aVNpW', 'https://watcha.com/ko-KR/contents/m5GwDD5', 'https://watcha.com/ko-KR/contents/mdE96Vd', 'https://watcha.com/ko-KR/contents/mOglQMd', 'https://watcha.com/ko-KR/contents/m5DGX1d', 'https://watcha.com/ko-KR/contents/mWyJYnx', 'https://watcha.com/ko-KR/contents/mWyJpxy', 'https://watcha.com/ko-KR/contents/mWpmnxO', 'https://watcha.com/ko-KR/contents/mdKB79y', 'https://watcha.com/ko-KR/contents/mWvkD0D', 'https://watcha.com/ko-KR/contents/mOlgPnW', 'https://watcha.com/ko-KR/contents/md76Pl8', 'https://watcha.com/ko-KR/contents/mdjal9O', 'https://watcha.com/ko-KR/contents/mO2MvNW', 'https://watcha.com/ko-KR/contents/m5mpXAd', 'https://watcha.com/ko-KR/contents/mOPoaXW', 'https://watcha.com/ko-KR/contents/mOk66kW', 'https://watcha.com/ko-KR/contents/mr5ZGLW', 'https://watcha.com/ko-KR/contents/m5Q4mGd', 'https://watcha.com/ko-KR/contents/m53MkyO', 'https://watcha.com/ko-KR/contents/mWzVYBW', 'https://watcha.com/ko-KR/contents/m5mQaj5', 'https://watcha.com/ko-KR/contents/mWyJ1v6', 'https://watcha.com/ko-KR/contents/md7ZBBO', 'https://watcha.com/ko-KR/contents/m5nQlkd', 'https://watcha.com/ko-KR/contents/mP5mD3d', 'https://watcha.com/ko-KR/contents/mdM3emW', 'https://watcha.com/ko-KR/contents/mdBzn2q', 'https://watcha.com/ko-KR/contents/mAd7yPW', 'https://watcha.com/ko-KR/contents/mWqn7XO', 'https://watcha.com/ko-KR/contents/mVWz0Pd', 'https://watcha.com/ko-KR/contents/m5rokDW', 'https://watcha.com/ko-KR/contents/mOPBqw5', 'https://watcha.com/ko-KR/contents/m533Lw5', 'https://watcha.com/ko-KR/contents/mOA9xlO', 'https://watcha.com/ko-KR/contents/mWpbEAd', 'https://watcha.com/ko-KR/contents/mdj1aJO', 'https://watcha.com/ko-KR/contents/mNdKPR5', 'https://watcha.com/ko-KR/contents/mdRKAD5', 'https://watcha.com/ko-KR/contents/mO0gZ8X', 'https://watcha.com/ko-KR/contents/m5X3JLd', 'https://watcha.com/ko-KR/contents/mWz6MBW', 'https://watcha.com/ko-KR/contents/mWpmmYK', 'https://watcha.com/ko-KR/contents/m5D2MDO', 'https://watcha.com/ko-KR/contents/mO2aDaW', 'https://watcha.com/ko-KR/contents/mOgV7AW', 'https://watcha.com/ko-KR/contents/m5xYKqW', 'https://watcha.com/ko-KR/contents/mdR0brW', 'https://watcha.com/ko-KR/contents/mdBzma9', 'https://watcha.com/ko-KR/contents/mDWvrxW', 'https://watcha.com/ko-KR/contents/mdEw9nb', 'https://watcha.com/ko-KR/contents/myWqkld', 'https://watcha.com/ko-KR/contents/mdMgLRW', 'https://watcha.com/ko-KR/contents/m5Xxqp5', 'https://watcha.com/ko-KR/contents/mAd7avW', 'https://watcha.com/ko-KR/contents/mP5mQx5', 'https://watcha.com/ko-KR/contents/md7q6MW', 'https://watcha.com/ko-KR/contents/mQO8YGd', 'https://watcha.com/ko-KR/contents/mOk6qjW', 'https://watcha.com/ko-KR/contents/mwWLKz5', 'https://watcha.com/ko-KR/contents/mdRLrvW', 'https://watcha.com/ko-KR/contents/m2djqJO', 'https://watcha.com/ko-KR/contents/mWJyv45', 'https://watcha.com/ko-KR/contents/mnWJ2QW', 'https://watcha.com/ko-KR/contents/mdja6rO', 'https://watcha.com/ko-KR/contents/mBOkqE5', 'https://watcha.com/ko-KR/contents/mLOPkXW', 'https://watcha.com/ko-KR/contents/my5YGwO', 'https://watcha.com/ko-KR/contents/m45ngDW', 'https://watcha.com/ko-KR/contents/m1WwXlO', 'https://watcha.com/ko-KR/contents/mP5mPzd', 'https://watcha.com/ko-KR/contents/mr5ZNK5', 'https://watcha.com/ko-KR/contents/mO8Zj35', 'https://watcha.com/ko-KR/contents/mO0glzW', 'https://watcha.com/ko-KR/contents/mdKB7PJ', 'https://watcha.com/ko-KR/contents/mOAR0Pd', 'https://watcha.com/ko-KR/contents/m5XBN4O', 'https://watcha.com/ko-KR/contents/md6YZaO', 'https://watcha.com/ko-KR/contents/mdEAKAd', 'https://watcha.com/ko-KR/contents/mW4alwO', 'https://watcha.com/ko-KR/contents/m5X3Rgd', 'https://watcha.com/ko-KR/contents/mWJy3J5', 'https://watcha.com/ko-KR/contents/mW4AagO', 'https://watcha.com/ko-KR/contents/mdBYBYO', 'https://watcha.com/ko-KR/contents/m5ejMyW', 'https://watcha.com/ko-KR/contents/mY5Q0Jd', 'https://watcha.com/ko-KR/contents/mOAZkQd', 'https://watcha.com/ko-KR/contents/m5nQMnd', 'https://watcha.com/ko-KR/contents/mgOARjd', 'https://watcha.com/ko-KR/contents/m5xJqxO', 'https://watcha.com/ko-KR/contents/myWqyBW', 'https://watcha.com/ko-KR/contents/mWpmAMO', 'https://watcha.com/ko-KR/contents/mM5rmEd', 'https://watcha.com/ko-KR/contents/mWJvZRW', 'https://watcha.com/ko-KR/contents/mMO2mNW', 'https://watcha.com/ko-KR/contents/mOoyZZ5', 'https://watcha.com/ko-KR/contents/mWwBQl5', 'https://watcha.com/ko-KR/contents/m5xMKrW', 'https://watcha.com/ko-KR/contents/m5QjklW', 'https://watcha.com/ko-KR/contents/m5ZxBRO', 'https://watcha.com/ko-KR/contents/m5eJP3W', 'https://watcha.com/ko-KR/contents/mVWzX9d', 'https://watcha.com/ko-KR/contents/mO00RLO', 'https://watcha.com/ko-KR/contents/md7ZPBO', 'https://watcha.com/ko-KR/contents/mWz4Z0d', 'https://watcha.com/ko-KR/contents/mXOgm05', 'https://watcha.com/ko-KR/contents/mOPo3VW', 'https://watcha.com/ko-KR/contents/mdEjNm5', 'https://watcha.com/ko-KR/contents/mW93M4d', 'https://watcha.com/ko-KR/contents/mmObLL5', 'https://watcha.com/ko-KR/contents/mr5Z8RO', 'https://watcha.com/ko-KR/contents/md6Y82O', 'https://watcha.com/ko-KR/contents/m5ZXwLO', 'https://watcha.com/ko-KR/contents/mE5DzQO', 'https://watcha.com/ko-KR/contents/mdR7xv5', 'https://watcha.com/ko-KR/contents/mmOb3kd', 'https://watcha.com/ko-KR/contents/m5XMeBW', 'https://watcha.com/ko-KR/contents/m5eQyz5', 'https://watcha.com/ko-KR/contents/mOAD1EW', 'https://watcha.com/ko-KR/contents/m5NvJX5', 'https://watcha.com/ko-KR/contents/mJOVXR5', 'https://watcha.com/ko-KR/contents/m5m14z5', 'https://watcha.com/ko-KR/contents/mQO8leO', 'https://watcha.com/ko-KR/contents/mQO8nX5', 'https://watcha.com/ko-KR/contents/mOo0Pzd', 'https://watcha.com/ko-KR/contents/mY5QVRW', 'https://watcha.com/ko-KR/contents/mdEww2W', 'https://watcha.com/ko-KR/contents/mLOPPPO', 'https://watcha.com/ko-KR/contents/m5Y6JvW', 'https://watcha.com/ko-KR/contents/m5xjRgO', 'https://watcha.com/ko-KR/contents/mO8jDq5', 'https://watcha.com/ko-KR/contents/mWyzmNO', 'https://watcha.com/ko-KR/contents/m5xQEqW', 'https://watcha.com/ko-KR/contents/mdjpxwW', 'https://watcha.com/ko-KR/contents/mLOPA7d', 'https://watcha.com/ko-KR/contents/m5X6o4O', 'https://watcha.com/ko-KR/contents/m5Y63jW', 'https://watcha.com/ko-KR/contents/mVWzB0W', 'https://watcha.com/ko-KR/contents/mOPopwW', 'https://watcha.com/ko-KR/contents/mWvkREO', 'https://watcha.com/ko-KR/contents/mObMNNW', 'https://watcha.com/ko-KR/contents/mWpkJw5', 'https://watcha.com/ko-KR/contents/mdKg2Rd', 'https://watcha.com/ko-KR/contents/m5Nla2O', 'https://watcha.com/ko-KR/contents/mdjlyzW', 'https://watcha.com/ko-KR/contents/m6dR6gd', 'https://watcha.com/ko-KR/contents/mQO8XVO', 'https://watcha.com/ko-KR/contents/m5Zo00W', 'https://watcha.com/ko-KR/contents/mYdE0OX', 'https://watcha.com/ko-KR/contents/mOAwoP5', 'https://watcha.com/ko-KR/contents/mYdE0EO', 'https://watcha.com/ko-KR/contents/mJOVBZO', 'https://watcha.com/ko-KR/contents/mO0mxad', 'https://watcha.com/ko-KR/contents/mOAVQZd', 'https://watcha.com/ko-KR/contents/m5awavO', 'https://watcha.com/ko-KR/contents/mW9rPKO', 'https://watcha.com/ko-KR/contents/mOgmjz5', 'https://watcha.com/ko-KR/contents/m5GwnB5', 'https://watcha.com/ko-KR/contents/mWp1EmO', 'https://watcha.com/ko-KR/contents/mO1QooO', 'https://watcha.com/ko-KR/contents/m5rgjEW', 'https://watcha.com/ko-KR/contents/m5ewEZO', 'https://watcha.com/ko-KR/contents/mObGNRd', 'https://watcha.com/ko-KR/contents/mn5NYnd', 'https://watcha.com/ko-KR/contents/mOV6XzO', 'https://watcha.com/ko-KR/contents/m5QbDRW', 'https://watcha.com/ko-KR/contents/mO2rjNO', 'https://watcha.com/ko-KR/contents/mOAw9G5', 'https://watcha.com/ko-KR/contents/mOkE1g5', 'https://watcha.com/ko-KR/contents/mWv1lb5', 'https://watcha.com/ko-KR/contents/mO2N00O', 'https://watcha.com/ko-KR/contents/mOVgp1d', 'https://watcha.com/ko-KR/contents/m5GvRKd', 'https://watcha.com/ko-KR/contents/mOAgnP5', 'https://watcha.com/ko-KR/contents/m5Dk2Rd', 'https://watcha.com/ko-KR/contents/mW4LMvL', 'https://watcha.com/ko-KR/contents/m5xY3qW', 'https://watcha.com/ko-KR/contents/mWpme6g', 'https://watcha.com/ko-KR/contents/mdMggKy', 'https://watcha.com/ko-KR/contents/mOPoEMa', 'https://watcha.com/ko-KR/contents/mdMgAmm', 'https://watcha.com/ko-KR/contents/mWzwo84', 'https://watcha.com/ko-KR/contents/m5nQxkn', 'https://watcha.com/ko-KR/contents/mWyQz6W', 'https://watcha.com/ko-KR/contents/m5ZXvVK', 'https://watcha.com/ko-KR/contents/m5mGVzO', 'https://watcha.com/ko-KR/contents/mWJyZ83', 'https://watcha.com/ko-KR/contents/m5rLzEO', 'https://watcha.com/ko-KR/contents/m5eJBq9', 'https://watcha.com/ko-KR/contents/mWzM9D5', 'https://watcha.com/ko-KR/contents/mQO8LqO', 'https://watcha.com/ko-KR/contents/mOAZxQd', 'https://watcha.com/ko-KR/contents/mYdEnE5', 'https://watcha.com/ko-KR/contents/mWLar6O', 'https://watcha.com/ko-KR/contents/mOPpNlO', 'https://watcha.com/ko-KR/contents/m5ZXrxO', 'https://watcha.com/ko-KR/contents/mOAPbEW', 'https://watcha.com/ko-KR/contents/mOgmjY5', 'https://watcha.com/ko-KR/contents/m5e8v3d', 'https://watcha.com/ko-KR/contents/mO2x9k5', 'https://watcha.com/ko-KR/contents/mdE09AO', 'https://watcha.com/ko-KR/contents/mW4L2XW', 'https://watcha.com/ko-KR/contents/m5rQoRd', 'https://watcha.com/ko-KR/contents/mOPooYW', 'https://watcha.com/ko-KR/contents/m5mQ9z5', 'https://watcha.com/ko-KR/contents/m5xNMjO', 'https://watcha.com/ko-KR/contents/m5NVoR5', 'https://watcha.com/ko-KR/contents/mdKmrJW', 'https://watcha.com/ko-KR/contents/mBOkBwW', 'https://watcha.com/ko-KR/contents/mjO0k7W', 'https://watcha.com/ko-KR/contents/m5nEZqd', 'https://watcha.com/ko-KR/contents/md76JPO', 'https://watcha.com/ko-KR/contents/m5n80DO', 'https://watcha.com/ko-KR/contents/m5NEynd', 'https://watcha.com/ko-KR/contents/m6dRLLW', 'https://watcha.com/ko-KR/contents/mW98ZM5', 'https://watcha.com/ko-KR/contents/mW4XgqW', 'https://watcha.com/ko-KR/contents/m5XXNL5', 'https://watcha.com/ko-KR/contents/m85X6BO', 'https://watcha.com/ko-KR/contents/mdja2NO', 'https://watcha.com/ko-KR/contents/mdMgZkW', 'https://watcha.com/ko-KR/contents/m533kl5', 'https://watcha.com/ko-KR/contents/mJ5x8x5', 'https://watcha.com/ko-KR/contents/m5e77oO', 'https://watcha.com/ko-KR/contents/mP5mgmO', 'https://watcha.com/ko-KR/contents/mW9GGZd', 'https://watcha.com/ko-KR/contents/mAd7k2d', 'https://watcha.com/ko-KR/contents/mWvkKMO', 'https://watcha.com/ko-KR/contents/mDWv6bW', 'https://watcha.com/ko-KR/contents/mW9208O', 'https://watcha.com/ko-KR/contents/m5rND75', 'https://watcha.com/ko-KR/contents/mWz4g4d', 'https://watcha.com/ko-KR/contents/m6dRLWy', 'https://watcha.com/ko-KR/contents/m15G9NO', 'https://watcha.com/ko-KR/contents/m2dj6Jd', 'https://watcha.com/ko-KR/contents/mOP73PW', 'https://watcha.com/ko-KR/contents/m5aVQDW', 'https://watcha.com/ko-KR/contents/md6xaaW', 'https://watcha.com/ko-KR/contents/m5Gvy2d', 'https://watcha.com/ko-KR/contents/mnWJ1n5', 'https://watcha.com/ko-KR/contents/mVWzvZW', 'https://watcha.com/ko-KR/contents/mWvDGZd', 'https://watcha.com/ko-KR/contents/mgOA3E5', 'https://watcha.com/ko-KR/contents/m5nGp9W', 'https://watcha.com/ko-KR/contents/mjO0leO', 'https://watcha.com/ko-KR/contents/mWpJoad', 'https://watcha.com/ko-KR/contents/mJ5xvEO', 'https://watcha.com/ko-KR/contents/mWqJBpO', 'https://watcha.com/ko-KR/contents/md7662O']

data_watcha=pd.DataFrame(columns=['id','title', 'title_ko','rate', 'year','country','genre','running_time','contents'])
watcha_review=pd.DataFrame(columns=['movie_id','user','rate','likes','review'])

chrome_options = webdriver.ChromeOptions()
chrome_options.add_argument('headless')
chrome_options.add_argument('--disable-gpu')
chrome_options.add_argument('lang=ko_KR')

for i in tqdm(range(0,len(movies))):
    #movie
    page=requests.get(movies[i])
    html=page.text
    soup=BeautifulSoup(html, 'html.parser')
    title_ko=soup.select('h1.css-13a04pq-Title.e1svyhwg14')[0].text
    rate_raw=soup.select('span.css-1r45xsg-RatingStatHeaderTopRight.e1dk30cn0')[0].text
    rate=rate_raw[rate_raw.find("★")+1:rate_raw.find("(")]
    
    #movie overview
    page=requests.get(movies[i]+"/overview")
    html=page.text
    soup=BeautifulSoup(html, 'html.parser')
    overview=soup.select('dl dd')
    
    title = overview[0].text
    year = overview[1].text
    country = overview[2].text
    genre = overview[3].text
    running_time = overview[4].text
    contents = overview[5].text

    #movie comment
    driver = webdriver.Chrome('./chromedriver', options=chrome_options)
    driver.implicitly_wait(3)

    driver.get(movies[i]+"/comments")

    SCROLL_PAUSE_TIME=2.5
    last_height = driver.execute_script("return document.body.scrollHeight")

    scroll_num = 0

    while True:
        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")

        time.sleep(SCROLL_PAUSE_TIME)
        driver.execute_script("window.scrollTo(0, document.body.scrollHeight-50);")
        time.sleep(SCROLL_PAUSE_TIME)

        new_height = driver.execute_script("return document.body.scrollHeight")

        if scroll_num == 6:
            break

        if new_height == last_height:
            break

        scroll_num += 1
        last_height = new_height

    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')

    review_list=soup.select('div.css-wnwcvo-Comment.e1oskw6f0')
    for review in review_list:
        try:
            user=review.select('a.css-1f9m1s4-StylelessLocalLink.e9fw3ul1')[0].text.replace(" ","")
            rate=review.select('div.css-1eyufz5-UserActionStatus.e1oskw6f4 span')[0].text
            try:
                likes=review.select('div.css-olvdbz-LikeReplyBlock.e1xxz10x1 em')[0].text
            except:
                likes=0.0
            try:
                comment=review.select('div.css-o67ghx-Text-handleRenderInner.e1xxz10x0')[0].text
            except:
                comment=review.select('div.css-1n96fhq-Text-handleRenderInner.e1xxz10x0')[0].text
            temp=pd.Series([i,user,rate,likes,comment],index=['movie_id','user','rate','likes','review'])
            watcha_review=watcha_review.append(temp,ignore_index=True)
        except:
            pass
        
    driver.close()
    
    temp=pd.Series([i,title,title_ko,rate,year,country,genre,running_time,contents],
                   index=['id','title', 'title_ko','rate', 'year','country','genre','running_time','contents'])
    data_watcha=data_watcha.append(temp,ignore_index=True)

data_watcha.to_csv('data_watcha.csv',encoding='utf-8-sig')
watcha_review.to_csv('watcha_review.csv', encoding='utf-8-sig')

data_watcha = pd.read_csv('data_watcha.csv', index_col=0)
watcha_review = pd.read_csv('watcha_review.csv', index_col=0)

각 영화에서 영화의 정보/장르/리뷰를 가져온다

data_watcha

watcha_review
