==> source/Watch/code/4. making_matrix.ipynb <==
import pandas as pd

# NCF data

file=pd.read_csv("../data/movie_watcha_review.csv",index_col=0)
file.head()

data={
    'u_id':file['user'],
    'p_id':file['movie_id'],
    'rating':file['rate']
}
score_board=pd.DataFrame(data, columns=['u_id','p_id','rating'])
score_board.sort_values(by=['u_id'], axis=0,inplace=True)
score_board.reset_index(drop = True, inplace=True)
score_board

score_board.to_csv('../data/score_board.csv', index=False)
