==> source/Watch/code/7. Watch_CBF.ipynb <==
import pandas as pd
import numpy as np
import math

# get size of vector
def mag(x): return math.sqrt(sum(i**2 for i in x))

movie_matrix = pd.read_csv('../data/movie_matrix.csv')
user_matrix = pd.read_csv('../data/user_adj_matrix.csv')
group_count = pd.read_csv('../data/group_count.csv',names=['group','count'])

movie_matrix.rename(columns = {"Unnamed: 0": "p_id"}, inplace=True)
movie_matrix.set_index('p_id',drop=True,inplace=True)
user_matrix.rename(columns = {"Unnamed: 0": "u_id_raw"}, inplace=True)
user_matrix.loc[user_matrix['u_id']=='0','u_id'] = user_matrix[user_matrix['u_id']=='0']['u_id_raw']
user_matrix.set_index('u_id', drop=True, inplace=True)
del user_matrix['u_id_raw']
group_count.set_index('group', drop=True, inplace=True)

user_matrix

movie_matrix

group_count.head()

p_m = pd.DataFrame(np.array(movie_matrix) * (1/np.array(group_count['count'])), index=movie_matrix.index)
p_m = p_m.iloc[:,:].div(p_m.sum(axis=1), axis=0).fillna(0)
u_m = pd.DataFrame(np.array(user_matrix) * (1/np.array(group_count['count'])), index=user_matrix.index)
u_m = u_m.iloc[:,:].div(u_m.sum(axis=1), axis=0).fillna(0)

p_m

u_m

p_m.to_csv('../data/normalized_item_matrix.csv')
u_m.to_csv('../data/normalized_user_matrix.csv')
